refactor(server): extract agent_builder_types to break builder cycle
The agent_builder <-> config/skill_builder <-> config/mcp_builder <->
ide/helpers cycle existed because config builders needed type imports
from agent_builder which transitively imported them back.

Extract all Pydantic models (AgentManifest, AgentFile, etc.) to a
standalone agent_builder_types.py module. Config builders now import
types from there, breaking the circular dependency.
